Virtual Machine/requirements.txt
boto==2.44.0
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
distro==1.0.1
firebase-admin==2.14.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-compute-engine==2.8.12
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
msgpack==0.6.0
paho-mqtt==1.4.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
python-apt==1.4.0b3
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.10.0
unattended-upgrades==0.1
urllib3==1.24.1